requirements.txt
# -- Core AI & LangChain Stack --
# These are the main libraries for your RAG application and guardrails.
langchain==0.3.27
langchain-core==0.3.76
langchain-community==0.3.29
nemoguardrails==0.18.0

# -- LLM & Embedding Model Integrations --
# Specific packages to connect to services like Google, Groq, and Hugging Face.
langchain-google-genai==2.1.12
langchain-groq==0.3.7
langchain-huggingface==0.3.1
sentence-transformers # A dependency for langchain-huggingface, good to list explicitly.

# -- Vector Store --
# The library for creating and managing your vector database.
faiss-cpu==1.12.0

# -- Development & Environment --
# Utilities for managing API keys and running code in notebooks.
python-dotenv==1.1.1
ipykernel==6.30.1
jupyter-client==8.6.3
jupyter-core==5.8.1

# -- Optional: Web API Framework --
# Uncomment these if you are building a web service with FastAPI.
# fastapi==0.121.1
# uvicorn==0.35.0

# -- Optional: Data Analysis --
# Useful libraries for data manipulation and analysis tasks.
# pandas==2.3.2
# numpy==1.26.4
sqlmodel